further re-organization of semantic diff
