+++ apollo-router/src/lib.rs
@@ -116,10 +116,10 @@ pub mod _private {
     pub use crate::plugin::PluginFactory;
     pub use crate::plugin::PLUGINS;
     // For comparison/fuzzing
-    pub use crate::query_planner::bridge_query_planner::render_diff;
     pub use crate::query_planner::bridge_query_planner::QueryPlanResult;
     pub use crate::query_planner::plan_compare::diff_plan;
     pub use crate::query_planner::plan_compare::plan_matches;
+    pub use crate::query_planner::plan_compare::render_diff;
     // For tests
     pub use crate::router_factory::create_test_service_factory_from_yaml;
 }
